+++ cmd/broadcaster-cli/app/utxos/create/create.go
@@ -51,7 +51,7 @@ var Cmd = &cobra.Command{
 			return err
 		}
 
-		miningFeeSat, err := helper.GetInt("miningFeeSatPerKb")
+		miningFeeSat, err := helper.GetUint64("miningFeeSatPerKb")
 		if err != nil {
 			return err
 		}
